+++ example/lib/entries/custom_context_menu_item.dart
@@ -22,7 +22,7 @@ final class CustomContextMenuItem extends ContextMenuItem<String> {
   }) : super.submenu();
 
   @override
-  bool get isFocusMaintained => true;
+  bool get autoHandleFocus => false;
 
   @override
   Widget builder(BuildContext context, ContextMenuState menuState,